Refactor: Migrate from deprecated `registerWith`
Removes the deprecated `registerWith` method in favor of the newer `onAttachedToEngine` for plugin registration. This also involves updating the `MethodChannel` instantiation to use `binaryMessenger` instead of the deprecated `getFlutterEngine().getDartExecutor()`.

Additionally, the code has been reformatted for better readability, and the `if` statement in `onMethodCall` has been converted to a `when` expression.